Weight converter between kgs and lbs
Using Python to create a simple weight converter based on the users input.
The user need to input their weight, and are then asked if they entered it in lbs or kgs. If it is entered in lbs, it gets converted to kgs, and if it is inserted in kgs, it will get converted to lbs.